Move the JSON indexer into the JSONField file
