+++ djangae/db/backends/appengine/indexing.py
@@ -780,54 +780,7 @@ class IRegexIndexer(RegexIndexer):
         return "_idx_iregex_{0}_{1}".format(field_column, self.get_pattern(index).encode('hex'))
 
 
-class JSONKeyLookupIndexer(Indexer):
-    OPERATOR = 'json_path'
 
-    def handles(self, field, operator):
-        from djangae.fields import JSONField
-
-        operator_part = operator.split("__", 1)[0]
-        return isinstance(field, JSONField) and operator_part == self.OPERATOR
-
-    def prepare_index_type(self, index_type, value):
-        return index_type
-
-    def prep_value_for_query(self, value):
-        return value
-
-    def prep_query_operator(self, op):
-        return "exact"
-
-    def prep_value_for_database(self, value, index):
-        import json
-
-        if isinstance(value, basestring):
-            try:
-                value = json.loads(value)
-            except ValueError:
-                return None
-
-        index_part = index.split("__", 1)[1]
-        path = index_part.split("__")
-
-        for section in path:
-            try:
-                section = int(section)
-            except (TypeError, ValueError):
-                pass
-
-            try:
-                value = value[section]
-            except (KeyError, IndexError, TypeError):
-                return None
-
-        return value
-
-    def indexed_column_name(self, field_column, value, index):
-        return "_idx_json_path_{}_{}".format(
-            field_column,
-            index.split("__", 1)[-1]
-        )
 
 
 _REGISTERED_INDEXERS = []
@@ -862,5 +815,3 @@ register_indexer(IStartsWithIndexer)
 register_indexer(RegexIndexer)
 register_indexer(IRegexIndexer)
 
-# Especially for JSON fields
-register_indexer(JSONKeyLookupIndexer)
